Add files via upload
First initial release of this script. It is a work in progress, but the general idea is solid and it should make quick work of resetting a plethora of browsers in the future. I also have an idea of making the script check which browsers are on the device and then give the user a list of available browsers to reset. Again this is a beta at best. Not sure how it will act on all systems and with all the possible browsers, only tested it on two on an Ubuntu system.