Add files via upload
Added additional information snippets to the Help. Added more info to certain functions in the script. Added an invalid number option which takes the user back to the previous menu on Install and Conquer. Added a couple of titles to the InstallandConquer function. Added a couple of more subdirectories to clean in Cleanup.