Add files via upload
Added option to scan for UFW in Arch if it's not there, switch automatically to enabling iptables firewall. Added optional kernel parameters to both to stop outside ping requests. Added the ability to remove old config files in Ubuntu toolbox and added possible solution to remove old kernels... This one is experimental, so if I was you and was at all unsure, I wouldn't do it. Added PlayonLinux and Wine to Install and Conquer in the Ubuntu script. Other minor changes and bugfixes.