Major change
Added a new function to create a swap file for those who need this. Occasionally, Manjaro and other distros when creating partitions will forget to create a swap partition, so this helps make it easy for new users to create one if they wish. There is a debate as to why swap is so important, but let's just say, you need some swap for the  suspend feature to work. I found some of this online, will cite the sources later.