New Beta
The all-new hosts updater, this time fully automatic. Can be ran from a root user cron job and  can be ran from cron.daily. This script uses fewer hosts right now which is also a good thing as it reduces the chance of false-positives and also accepts flags. Flags are a way to pass arguments to a bash script as it is being ran. This can help where choosing a certain function for a certain job is necessary. In this case, the hosts file you want to use is directly proportional with the Lettered flag you select at run time. If you're putting this in a root cron job, it's recommended that you write the script name and then the flags as in: ./hostsupdate -ABC. This will result in the first three hosts files being chosen and concatenated together.