Trivial bugfix
Some people use all 0's in the hopes that it will spare them that fraction of a second it takes the host to timeout during ping. This is nice when you have an especially large file, however, on most Unix systems, users would hardly notice a significant difference. The only time that would remotely be an issue is if you had an actual dns server running in the background on the same device on which you installed the hosts file to begin with. This is a problem for Windows, not so much in Linux.