Update to Hostsman4linux
Removed links to files that are either no longer curated or aren't being frequently updated and added Hphosts back. On the hphosts website, it is frowned upon to use the entire file as there may be false positives, has been deprecated, however, I believe that it is the most comprehensive file to use when torrenting these days.