Add files via upload
As is no longer necessary have removed some useless features such as removing hashes etc and have improved on the final look of the hosts file and added solid sources to be mirrored here at least for the time being. To add more domains see blacklist.txt. To add more hosts files or review the ones already listed see sources.txt. Tweaked checking for hosts.bak and how it applies the backup file in place of the original in the event that it exists. Previously this used a while loop and then did a separate operation to merge the backup file with the original, this is now all done as one operation which should make it more seamless. Readded some hosts files from back before. Removed an unnecessary warning label as it is just not needed with the dialogue at the beginning of the script.