Create 10-monitor.conf
This is an example xorg config monitor file to maintain screen resolution settings. This can also help maintain dual monitor settings, but this file will need to be tweaked for that. I already tweaked it on my system hence the 1440x900 setting. I got the template from Chris Titus on his youtube channel. This is good when xorg updates as it sometimes changes your screen resolution at least in arch linux.