Add files via upload
Added the optional Memory reallocation to GPU which will help in some video playback and graphical intensive situations, however, this has to be uncommented out to work as 
not all systems have enough RAM to spare. The GPU is integrated, so this shouldn't be a problem anyway, as it only uses memory if available and only enough to do the work 
in the first place, however, this should be enabled with caution and a bit of know how to disable it should things go wrong.