Add files via upload
Added a line of code which when enabled will set up cpu overclock to 1800 Mhz or 1.8 Ghz which is a slight step up from the current 1.5 Ghz max with most pi 4 model boards. Only use this on pi 4's and works by default on 8GB models but some 4GB models can benefit from this too.